Added changes for 1.9:
This is the biggest set of changes to the JSONDeserializer since it was written.  This
refactoring makes the deserializer much more flexible in responding to alternate
serialization formats using ObjectFactory.  ObjectFactories are the basis for all
type binding between JSON and Java Objects.  ObjectFactories can be configured by
path or Class.  Paths only apply that ObjectFactory to the object graph path.  Class
apply to all Class types detected in that object graph. 




git-svn-id: svn://svn.code.sf.net/p/flexjson/code/trunk@57 1649a90f-f54c-0410-9814-d0bc849eed40